requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

asgiref==3.4.1
azure-functions==1.7.2
click==8.0.1
fastapi==0.68.0
h11==0.12.0
pydantic==1.8.2
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
